+++ src/layers/PickDifficultyLayer.hpp
@@ -11,19 +11,19 @@ private:
     CCLabelBMFont* m_listNameLabel = nullptr;
 
     Function<void(Result<>)> m_loadedCallback;
-    Function<void(int, Difficulty, std::vector<LevelRanking>)> m_selectedCallback;
+    Function<void(int64_t, Difficulty, std::vector<LevelRanking>)> m_selectedCallback;
 
     std::vector<UserList> m_lists;
 
     std::vector<DifficultyNode*> m_difficultyNodes;
 
-    int m_levelID;
+    int64_t m_levelID;
     int m_listIndex = 0;
 
     bool m_isLoading = true;
     bool m_isSwitching = false;
 
-    PickDifficultyLayer(int, Function<void(Result<>)>, Function<void(int, Difficulty, std::vector<LevelRanking>)>);
+    PickDifficultyLayer(int64_t, Function<void(Result<>)>, Function<void(int64_t, Difficulty, std::vector<LevelRanking>)>);
 
     bool init() override;
 
@@ -35,6 +35,6 @@ private:
 
 public:
 
-    static PickDifficultyLayer* create(int, Function<void(Result<>)>, Function<void(int, Difficulty, std::vector<LevelRanking>)>);
+    static PickDifficultyLayer* create(int64_t, Function<void(Result<>)>, Function<void(int64_t, Difficulty, std::vector<LevelRanking>)>);
 
 };
